+++ TUPScheduling/templates/class_schedule/class_schedule.html
@@ -72,7 +72,7 @@
                   <h1 class="font-bold">TIME</h1>
                 </div>
               </a>
-              <a href="/class-schedule-overview">
+              <a href="class-schedule-overview">
                 <div class="px-3 py-3 bg-white border-4 border-gray-400 rounded-r-2xl w-28">
                   <h1 class="font-bold">OVERVIEW</h1>
                 </div>
