+++ TUPScheduling/templates/class_schedule/class_schedule_overview.html
@@ -69,7 +69,7 @@
                 <h1 class="font-bold">TIME</h1>
                 </div>
             </a>
-            <a href="/class-schedule-overview">
+            <a href="/">
                 <div class="px-3 py-3 bg-gray-400 border-4 border-gray-400 rounded-r-2xl w-28">
                 <h1 class="font-bold">OVERVIEW</h1>
                 </div>
