+++ TUPScheduling/urls.py
@@ -6,7 +6,7 @@ from wagtail.admin import urls as wagtailadmin_urls
 from wagtail.core import urls as wagtail_urls
 from wagtail.documents import urls as wagtaildocs_urls
 
-from search import views as search_views
+from TUPScheduling.search import views as search_views
 
 urlpatterns = [
     path('django-admin/', admin.site.urls),
