--- administrator/views.py
@@ -1,4 +1,0 @@
-
-from wagtail.contrib.modeladmin.views import CreateView
-
-## if ever need natin ng views
